fet: Docer запускает пустой сервер
