+++ docker-compose.yml
@@ -1,24 +1,23 @@
 version: '3.1'
 services:
   stats-server:
-    build: ./stats-server/stat-server
-    image: stats-server
-    tty: true
+    build: ./stats-server/server/
+    #image: stats-server
+    #tty: true
     container_name: stats-server
     ports:
       - "9090:9090"
-      - "9091:9091"
+     # - "9091:9091"
     depends_on:
       - stats-db
+      - ewm-service
     environment:
-      #- SPRING_PROFILES_ACTIVE=test
       - DB_HOST=stats-db
       - DB_PORT=5432
       - DB_NAME=stats-server
       - DB_USER=postgres
       - DB_PASSWORD=root
-      - SPRING_PROFILES_ACTIVE=default
-      #- SPRING_DATASOURCE_URL=jdbc:postgresql://stats-db:5432/stats-db
+      #- SPRING_PROFILES_ACTIVE=default
 
   stats-db:
     image: postgres:14-alpine
@@ -31,24 +30,22 @@ services:
       - POSTGRES_PASSWORD=root
 
   ewm-service:
-    build: ewm-service
-    image: ewm-service
-    tty: true
+    build: ewm-service/
+    #image: ewm-service
+    #tty: true
     container_name: ewm-service
     ports:
       - "8080:8080"
-      - "8081:8081"
+      #- "8081:8081"
     depends_on:
       - ewm-db
-      - stats-server
+      #- stats-server
     environment:
       - DB_HOST=ewm-db
       - DB_PORT=5432
       - DB_NAME=ewm-service
       - DB_USER=postgres
       - DB_PASSWORD=root
-      #- APP_Stats_Server_URL=http://stats-server:9090
-      #- SERVER_PORT=8080
 
   ewm-db:
     image: postgres:14-alpine
