+++ ewm-service/Dockerfile
@@ -1,4 +1,5 @@
 FROM amazoncorretto:11-alpine-jdk
-ENV JAVA_TOOL_OPTIONS -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:8080
+#ENV JAVA_TOOL_OPTIONS -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:8080
+WORKDIR /app
 COPY target/*.jar app.jar
-ENTRYPOINT ["java","-jar","/EwmServise.jar"]
+ENTRYPOINT ["java","-jar","/app/app.jar"]
