+++ stats-server/client/src/main/java/ru/practicum/stats/StatClient.java
@@ -1,0 +1,11 @@
+package ru.practicum.stats;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatClient {
+    public static void main(String[] args) {
+        SpringApplication.run(StatClient.class, args);
+    }
+}
