+++ stats-server/client/src/main/resources/application.properties
