+++ stats-server/server/src/main/java/ru/practicum/stats/StatsServer.java
@@ -1,0 +1,11 @@
+package ru.practicum.stats;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatsServer {
+    public static void main(String[] args) {
+        SpringApplication.run(StatsServer.class, args);
+    }
+}
