+++ stats-server/server/src/main/java/ru/practicum/stats/hendler/ErrorResponse.java
@@ -1,0 +1,13 @@
+package ru.practicum.stats.hendler;
+
+public class ErrorResponse {
+    private final String error;
+
+    public ErrorResponse(String error) {
+        this.error = error;
+    }
+
+    public String getError() {
+        return error;
+    }
+}
