+++ stats-server/server/src/main/java/ru/practicum/stats/model/EndpointHit.java
@@ -1,0 +1,32 @@
+package ru.practicum.stats.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+
+@Data
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Table(name = "endpoint_hits")
+public class EndpointHit {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String app;
+
+    private String uri;
+
+    private String ip;
+
+    private LocalDateTime timestamp;
+
+}
