+++ stats-server/server/src/main/java/ru/practicum/stats/model/ViewStat.java
@@ -1,0 +1,19 @@
+package ru.practicum.stats.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ViewStat {
+
+    private String app;
+
+    private String uri;
+
+    private Long hits;
+}
