+++ stats-server/server/src/main/resources/application.properties
