+++ stats-server/server/src/main/resources/schema.sql
