--- stats-server/src/main/java/ru/practicum/StatsServer.java
@@ -1,11 +1,0 @@
-package ru.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatsServer {
-    public static void main(String[] args) {
-        SpringApplication.run(StatsServer.class, args);
-    }
-}
