--- stats-server/stat-client/src/main/java/ru/practicum/StatClient.java
@@ -1,11 +1,0 @@
-package ru.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatClient {
-    public static void main(String[] args) {
-        SpringApplication.run(StatClient.class, args);
-    }
-}
