--- stats-server/stat-server/Dockerfile
@@ -1,4 +1,0 @@
-FROM amazoncorretto:11-alpine-jdk
-ENV JAVA_TOOL_OPTIONS -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:9090
-COPY target/*.jar app.jar
-ENTRYPOINT ["java","-jar","/StatServer.jar"]
