--- stats-server/stat-server/src/main/java/ru/practicum/StatServer.java
@@ -1,11 +1,0 @@
-package ru.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatServer {
-    public static void main(String[] args) {
-        SpringApplication.run(StatServer.class, args);
-    }
-}
