--- stats-server/stat-server/src/main/java/ru/practicum/dto/ViewStatDto.java
@@ -1,18 +1,0 @@
-package ru.practicum.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ViewStatDto {
-    private String app;
-
-    private String uri;
-
-    private Long hits;
-}
