--- stats-server/stat-server/src/main/java/ru/practicum/hendler/ErrorResponse.java
@@ -1,13 +1,0 @@
-package ru.practicum.hendler;
-
-public class ErrorResponse {
-    private final String error;
-
-    public ErrorResponse(String error) {
-        this.error = error;
-    }
-
-    public String getError() {
-        return error;
-    }
-}
