--- stats-server/stat-server/src/main/java/ru/practicum/model/EndpointHit.java
@@ -1,32 +1,0 @@
-package ru.practicum.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-
-@Data
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Table(name = "endpoint_hits")
-public class EndpointHit {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String app;
-
-    private String uri;
-
-    private String ip;
-
-    private LocalDateTime timestamp;
-
-}
