--- stats-server/stat-server/src/main/java/ru/practicum/model/ViewStat.java
@@ -1,19 +1,0 @@
-package ru.practicum.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ViewStat {
-
-    private String app;
-
-    private String uri;
-
-    private Long hits;
-}
