--- stats-server/stat-server/src/main/java/ru/practicum/service/EndpointHitService.java
@@ -1,14 +1,0 @@
-package ru.practicum.service;
-
-
-import ru.practicum.dto.EndpointHitDto;
-import ru.practicum.dto.ViewStatDto;
-
-import java.util.List;
-
-public interface EndpointHitService {
-
-    EndpointHitDto save(EndpointHitDto endpointHitDto);
-
-    List<ViewStatDto> findAllViewStats(List<String> uris, String start, String end, Boolean unique);
-}
